requirements.txt
streamlit
pandas
pdfplumber
plotly
fpdf2
kaleido
openpyxl